Rename paths -> discovery
