+++ src/sqlfluff/core/linter/discovery.py
